+++ src/main/java/com/example/petclinic/service/OwnerService.java
@@ -23,7 +23,7 @@ public class OwnerService implements BasicService<Owner> {
     }
 
     @Override
-    public Owner get(int id) {
+    public Owner get(Long id) {
 
         return this.ownerRepository.read(new Owner(id));
     }
